move tcp tunnel code to separate file
